Fix some parameter type

+++ app/src/main/java/ch/epfl/sdp/drone3d/service/api/location/LocationService.kt
@@ -35,7 +35,7 @@ interface LocationService {
     fun subscribeToLocationUpdates(
         consumer: (LatLng) -> Unit,
         minTimeDelta: Long,
-        minDistanceDelta: Long
+        minDistanceDelta: Float
     ): Int?
 
     /**
